+++ src/web/assets/livepreview/src/BreakpointEditor.vue
@@ -1,0 +1,71 @@
+<template>
+  <div
+    :class="{ enabled, dragging }"
+    class="bp-editor"
+    @dragstart.native="e => console.log('DRAGSTART', e)"
+  >
+    <BreakpointButton />
+
+    <EditorToolbar v-if="enabled" />
+
+    <EditorStage>
+      <EditorDraggable v-if="enabled" />
+      <EditorViewport :iframe="iframe" />
+    </EditorStage>
+  </div>
+</template>
+
+<script>
+import { mapGetters } from "vuex";
+import BreakpointButton from "./BreakpointButton";
+import EditorDraggable from "./EditorDraggable";
+import EditorToolbar from "./EditorToolbar";
+import EditorStage from "./EditorStage";
+import EditorViewport from "./EditorViewport";
+
+export default {
+  name: "BreakpointEditor",
+
+  components: {
+    BreakpointButton,
+    EditorDraggable,
+    EditorStage,
+    EditorToolbar,
+    EditorViewport
+  },
+
+  props: {
+    iframe: {
+      type: Object,
+      required: true
+    }
+  },
+
+  computed: mapGetters(["enabled", "dragging"])
+};
+</script>
+
+<style scoped>
+.bp-editor {
+  position: relative;
+  width: 100%;
+  height: 100%;
+  background-color: #f7f7f7;
+  overflow: hidden;
+}
+
+.dragging .bp-editor {
+  pointer-events: none;
+}
+
+.bp-editor.enabled >>> .bp-stage {
+  top: 38px;
+  left: 20px;
+  right: 20px;
+  bottom: 20px;
+}
+
+.bp-editor.dragging >>> .bp-viewport {
+  pointer-events: none;
+}
+</style>